# backend/services/auth/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud-Shop Auth Service API
  description: Authentication and authorization service for Cloud-Shop
  version: 1.0.0
  contact:
    name: Cloud-Shop Development Team
    email: dev@cloud-shop.com

servers:
  - url: http://localhost:8081
    description: Development server

paths:
  /cloud-shop/auth/login:
    post:
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cloud-shop/auth/register:
    post:
      summary: User registration
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
              required:
                - email
                - password
                - name
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cloud-shop/auth/confirm:
    post:
      summary: Confirm user registration
      operationId: confirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
              required:
                - email
                - code
      responses:
        '200':
          description: Confirmation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid confirmation code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cloud-shop/auth/logout:
    post:
      summary: User logout
      operationId: logout
      requestBody:
        required: false
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cloud-shop/auth/verify:
    post:
      summary: Verify JWT token
      operationId: verifyToken
      requestBody:
        required: false
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cloud-shop/auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  expires_in:
                    type: integer
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cloud-shop/auth/revoke:
    post:
      summary: Revoke token
      operationId: revokeToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
              required:
                - token
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /cloud-shop/auth/users/{userId}/roles:
    get:
      summary: Get user roles
      operationId: getUserRoles
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User roles retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'

  /cloud-shop/auth/roles/assign:
    post:
      summary: Assign role to user
      operationId: assignRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                roleId:
                  type: integer
              required:
                - userId
                - roleId
      responses:
        '200':
          description: Role assigned successfully

  /cloud-shop/auth/roles/remove:
    post:
      summary: Remove role from user
      operationId: removeRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                roleId:
                  type: integer
              required:
                - userId
                - roleId
      responses:
        '200':
          description: Role removed successfully

components:
  schemas:
    User:
      type: object
      properties:
        sub:
          type: string
        email:
          type: string
        name:
          type: string
        preferred_username:
          type: string
        roles:
          type: array
          items:
            type: string

    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time